auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.13"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pymongo = "^4.6.0"
pydantic = "^2.5.0"
redis = "^5.0.1"
pytz = "^2025.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
